move channels into router.go
